requirements.txt
flask
black
flask-wtf
email-validator
flask-sqlalchemy
flask-bcrypt
flask-login
pillow